src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/attr-title-inherit-expected.txt

CONSOLE ERROR: line 49: Uncaught TypeError: host.createShadowRoot is not a function
This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught TypeError: host.createShadowRoot is not a function
Harness: the test ran to completion.
